+++ R/hypothesis_formula.R
@@ -9,6 +9,16 @@ hypothesis_formula_list <- list(
             label = function(x) sprintf("(%s) - (%s)", x, x[1])[2:length(x)]
         )
     ),
+    revreference = list(
+        ratio = list(
+            comparison = function(x) (x[1] / x)[2:length(x)],
+            label = function(x) sprintf("(%s) / (%s)", x[1], x)[2:length(x)]
+        ),
+        difference = list(
+            comparison = function(x) (x[1] - x)[2:length(x)],
+            label = function(x) sprintf("(%s) - (%s)", x[1], x)[2:length(x)]
+        )
+    ),
     sequential = list(
         ratio = list(
             comparison = function(x) (x / data.table::shift(x))[2:length(x)],