+++ NEWS.md
@@ -8,6 +8,7 @@ Bugs:
 * Error when `hypotheses()` has both a number in `hypothesis` and a value in `multcomp`. Thanks to J. Rohrer for Issue #1381.
 * Informative error: conformal inference is not supported for `tidymodels`
 * `type` is respected for models of class `betareg`. Thanks to @strengejacke for reporting issue #1391.
+* `revreference` is available again in the `hypothesis` argument.
 
 Minor: